requirements.txt
Django==1.5.1
South==0.7.6
argparse==1.2.1
dj-database-url==0.2.1
django-forms-bootstrap==2.0.3.post1
django-jsonfield==0.9.4
django-nose==1.1
django-stripe-payments==2.0b20
mock==1.0.1
nose==1.3.0
psycopg2==2.5
requests==1.2.1
stripe==1.7.9
wsgiref==0.1.2
